Update adding courses function
