+++ db.json
@@ -1,107 +1,115 @@
 {
   "users": [
-    { "id": "admin", "name": "Lord", "password": "12345" },
-    { "id": "pablo", "name": "Pablo", "password": "123456" },
-    { "id": "mike", "name": "Mike", "password": "123456" }
-  ],
-  "info": [
-    {"id": "users_length", "data": 3},
-    {"id": "courses_length", "data": 12}
+    {
+      "id": "admin",
+      "name": "Lord",
+      "password": "12345"
+    },
+    {
+      "id": "pablo",
+      "name": "Pablo",
+      "password": "123456"
+    },
+    {
+      "id": "mike",
+      "name": "Mike",
+      "password": "123456"
+    }
   ],
   "courses": [
     {
-      "id":1,
-      "title":"Angular 4222",
-      "description":"Angular 4222 course",
-      "authorID":"pablo",
+      "id": 1,
+      "title": "Angular 4222",
+      "description": "Angular 4222 course",
+      "authorID": "pablo",
       "date": "2020-04-16T10:55:55.058Z",
       "rating": 3
     },
     {
-      "id":2,
-      "title":"Angular.js",
-      "description":"Angular.js course",
-      "authorID":"mike",
+      "id": 2,
+      "title": "Angular.js",
+      "description": "Angular.js course",
+      "authorID": "mike",
       "date": "2021-04-16T10:55:55.058Z",
       "rating": 5
     },
     {
-      "id":3,
-      "title":"Angular 6",
-      "description":"Angular 6 course",
-      "authorID":"pablo",
+      "id": 3,
+      "title": "Angular 6",
+      "description": "Angular 6 course",
+      "authorID": "pablo",
       "date": "2020-05-16T10:55:55.058Z",
       "rating": 3
     },
     {
-      "id":4,
-      "title":"Angular 7",
-      "description":"Angular 7 course",
-      "authorID":"mike",
+      "id": 4,
+      "title": "Angular 7",
+      "description": "Angular 7 course",
+      "authorID": "mike",
       "date": "2019-04-16T10:55:55.058Z",
       "rating": 5
     },
     {
-      "id":5,
-      "title":"Angular 8",
-      "description":"Angular 8 course",
-      "authorID":"mike",
+      "id": 5,
+      "title": "Angular 8",
+      "description": "Angular 8 course",
+      "authorID": "mike",
       "date": "2018-04-16T10:55:55.058Z",
       "rating": 5
     },
     {
-      "id":6,
-      "title":"Angular 2",
-      "description":"Angular 2 course",
-      "authorID":"mike",
+      "id": 6,
+      "title": "Angular 2",
+      "description": "Angular 2 course",
+      "authorID": "mike",
       "date": "2018-04-16T10:55:55.058Z",
       "rating": 5
     },
     {
-      "id":7,
-      "title":"Angular 4",
-      "description":"Angular 4 course",
-      "authorID":"pablo",
+      "id": 7,
+      "title": "Angular 4",
+      "description": "Angular 4 course",
+      "authorID": "pablo",
       "date": "2020-04-16T10:55:55.058Z",
       "rating": 3
     },
     {
-      "id":8,
-      "title":"Angular.js",
-      "description":"Angular.js course",
-      "authorID":"mike",
+      "id": 8,
+      "title": "Angular.js",
+      "description": "Angular.js course",
+      "authorID": "mike",
       "date": "2021-04-16T10:55:55.058Z",
       "rating": 5
     },
     {
-      "id":9,
-      "title":"Angular 6",
-      "description":"Angular 6 course",
-      "authorID":"pablo",
+      "id": 9,
+      "title": "Angular 6",
+      "description": "Angular 6 course",
+      "authorID": "pablo",
       "date": "2020-05-16T10:55:55.058Z",
       "rating": 3
     },
     {
-      "id":10,
-      "title":"Angular 7",
-      "description":"Angular 7 course",
-      "authorID":"mike",
+      "id": 10,
+      "title": "Angular 7",
+      "description": "Angular 7 course",
+      "authorID": "mike",
       "date": "2019-04-16T10:55:55.058Z",
       "rating": 5
     },
     {
-      "id":11,
-      "title":"Angular 8",
-      "description":"Angular 8 course",
-      "authorID":"mike",
+      "id": 11,
+      "title": "Angular 8",
+      "description": "Angular 8 course",
+      "authorID": "mike",
       "date": "2018-04-16T10:55:55.058Z",
       "rating": 5
     },
     {
-      "id":12,
-      "title":"Angular 2",
-      "description":"Angular 2 course",
-      "authorID":"mike",
+      "id": 12,
+      "title": "Angular 2",
+      "description": "Angular 2 course",
+      "authorID": "mike",
       "date": "2018-04-16T10:55:55.058Z",
       "rating": 5
     }
