+++ src/Interface.ts
