+++ src/components/HomePage/Home.tsx
@@ -2,7 +2,7 @@
 import { useEffect, useState } from 'react';
 import { apiStates, useApi } from '../../tools';
 import Course from "../Course/Course";
-import {CourseModel} from "../../mockfileModel"
+import {CourseModel} from "../../Interface"
 
 const LIMIT_OF_ITEMS = 9;
